REFACT - Split overlong `model.py` module - divided into `model_sources.py` and `model_properties.py` as logical split.
